Extract reusable media preview component
