+++ app/build.gradle.kts
@@ -153,6 +153,8 @@ dependencies {
     implementation(libs.coil.video)
     implementation(libs.reorderable)
     implementation(libs.telephoto.zoomable.image.coil)
+    implementation(libs.media3.exoplayer)
+    implementation(libs.media3.ui)
 
     // Hilt
     implementation(libs.hilt.android)
